Added levelManager and things related to entities
LevelManager finished (must be tested),
Added Entity,Blocks and redBlock (must add physics, animation and logic
to be destroyed.),
Added some files related to levesl (such as Test_01 and the path to this
file in FilesToLoad.json). Also, the FilesManager can load the json
files of the levels (must be tested).